organize test functions and add a test for detecting third-order edges
